Moved interdiff function to `phabricator.py`
